+++ app/routers/__init__.py
@@ -1,0 +1,5 @@
+import app.routers.closet
+import app.routers.favicon
+import app.routers.randomize
+import app.routers.root
+import app.routers.show
