+++ pyproject.toml
@@ -9,6 +9,7 @@ python = "^3.10"
 fastapi = "^0.78.0"
 uvicorn = {extras = ["standard"], version = "^0.17.6"}
 opencv-python = "^4.6.0.66"
+requests = "^2.28.1"
 
 [tool.poetry.dev-dependencies]
 autopep8 = "^1.7.0"
