+++ requirements.txt
@@ -1,19 +1,23 @@
 anyio==3.6.1 ; python_version >= "3.10" and python_version < "4.0"
 asgiref==3.5.2 ; python_version >= "3.10" and python_version < "4.0"
+certifi==2022.9.24 ; python_version >= "3.10" and python_version < "4"
+charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
 click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
 colorama==0.4.5 ; sys_platform == "win32" and python_version >= "3.10" and python_version < "4.0" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
 fastapi==0.78.0 ; python_version >= "3.10" and python_version < "4.0"
 h11==0.13.0 ; python_version >= "3.10" and python_version < "4.0"
 httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
-idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
+idna==3.4 ; python_version >= "3.10" and python_version < "4"
 numpy==1.23.3 ; python_version >= "3.10" and python_version < "4.0"
 opencv-python==4.6.0.66 ; python_version >= "3.10" and python_version < "4.0"
 pydantic==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
 python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
 pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
+requests==2.28.1 ; python_version >= "3.10" and python_version < "4"
 sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
 starlette==0.19.1 ; python_version >= "3.10" and python_version < "4.0"
 typing-extensions==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
+urllib3==1.26.12 ; python_version >= "3.10" and python_version < "4"
 uvicorn[standard]==0.17.6 ; python_version >= "3.10" and python_version < "4.0"
 uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
 watchgod==0.8.2 ; python_version >= "3.10" and python_version < "4.0"
